Preferences now saves correct wiki config.json with admin
Wiki's admin must keep values in config.json in sync with owner.json.

We moved code back into preferences/window.js. It was previously in
src/index.js when we were debugging windowing behavior.

We also moved the electron-store code into ./preferences/*.

Peculiar bit of async code now required in preferences/preload.js.
config.json needs to specify the path to owner.json. That requires
access to the app's named path "userData". app is not available in the
preload.js script. So we need to setup a protocol for preload.js to
use ipcRenderer to ask ipcMain for that path. The call is async.