proof that wiki runs fine outside electron and fails inside electron
This commit holds a whole lot of frustration. I have reduced the
example as much as I can stand for the time being.

We learned a fair amount about how fork() in Electron differs from
fork() outside Electron from a blog post entitled:

Everything You Wanted To Know About Electron Child Processes.
https://www.matthewslipper.com/2019/09/22/everything-you-wanted-electron-child-process.html

We took some of the example code as a starting point for our current
experiment:

https://github.com/mslipper/electron-child-process-playground/blob/master/src/main.js#L56-L58

So... the symptom:

  works: npm run spike
  fails: npm run start

The path that works runs a tiny node app which forks() federated
wiki. In that path we can login and edit pages.

the path that fails runs an electron app which forks() the same
federated wiki program. In that path we cannot even load a page.

Fails when we try to visit a page with ERR_CRYPTO_UNKNOWN_CIPHER

The part of the electron app which works and which I care about is a
preferences panel which allows one to edit the author name and
password in those config files. I have gone to some lengths to make
the non-electron work with the config files in the same place on the
file system that electron-store uses.

Note: there is coupling through paths in the filesystem which are not
visible in the code. The code in config.js has to know about the
default behavior for electron-store.